ImplementacionCRUD-Sin Terminar

+++ SeguimientoNutricional.App/SeguimientoNutricional.App.Dominio/Entidades/Paciente.cs
@@ -6,9 +6,9 @@ namespace SeguimientoNutricional.App.Dominio
     {    
         //public int Id {get; set; }
 
-        public int Latitud {get; set; }
+        public float Latitud {get; set; }
 
-        public int Longitud {get; set; }   
+        public float Longitud {get; set; }   
 
         public string Direccion {get; set; }
